* StatsPrint round numbers in prints fix

+++ lpd/callbacks/callback_base.py
@@ -20,11 +20,11 @@ class CallbackBase():
         self.apply_on_states = apply_on_states
         self.round_values_on_print_to = round_values_on_print_to
 
-    def round_to(self, value: Union[int,list]):
+    def round_to(self, value: Union[float,list]):
         if self.round_values_on_print_to:
             if isinstance(value, list):
                 return [round(v, self.round_values_on_print_to) for v in value]
-            if isinstance(value, int):
+            if isinstance(value, float):
                 return round(value, self.round_values_on_print_to)
         return value
